lnwallet: move keyring to new commitmenttx package
